requirements.txt
streamlit
plotly
requests
python-dotenv
pandas
